cleanup, make logo as a prop
