+++ app/scripts/components/common/layout-root/index.tsx
@@ -3,7 +3,6 @@ import { useDeepCompareEffect } from 'use-deep-compare';
 import styled from 'styled-components';
 import { Outlet } from 'react-router';
 import { reveal } from '@devseed-ui/animation';
-import { getString, getNavItemsFromVedaConfig } from 'veda';
 import MetaTags from '../meta-tags';
 import PageFooter from '../page-footer';
 import Banner from '../banner';
@@ -12,16 +11,8 @@ import { LayoutRootContext } from './context';
 import { useGoogleTagManager } from '$utils/use-google-tag-manager';
 
 import NavWrapper from '$components/common/nav-wrapper';
-import { InternalNavLink, ExternalNavLink, ModalNavLink, DropdownNavLink } from '$components/common/page-header';
-
-import { checkEnvFlag } from '$utils/utils';
-import {
-  STORIES_PATH,
-  DATASETS_PATH,
-  ANALYSIS_PATH,
-  EXPLORATION_PATH,
-  ABOUT_PATH
-} from '$utils/routes';
+import Logo from '$components/common/page-header/logo';
+import { mainNavItems, subNavItems} from '$components/common/page-header/default-config';
 
 const appTitle = process.env.APP_TITLE;
 const appDescription = process.env.APP_DESCRIPTION;
@@ -43,44 +34,6 @@ const PageBody = styled.div`
   overflow-anchor: auto;
 `;
 
-
-let defaultMainNavItems:(ExternalNavLink | InternalNavLink | DropdownNavLink | ModalNavLink)[] = [{
-  title: 'Data Catalog',
-  to: DATASETS_PATH,
-  type: 'internalLink'
-}, {
-  title: checkEnvFlag(process.env.FEATURE_NEW_EXPLORATION) ? 'Exploration' : 'Analysis',
-  to: checkEnvFlag(process.env.FEATURE_NEW_EXPLORATION) ? EXPLORATION_PATH : ANALYSIS_PATH,
-  type: 'internalLink'
-}, {
-  title: getString('stories').other,
-  to: STORIES_PATH,
-  type: 'internalLink'
-}];
-
-if (!!process.env.HUB_URL && !!process.env.HUB_NAME) defaultMainNavItems = [...defaultMainNavItems, {
-  title: process.env.HUB_NAME,
-  href: process.env.HUB_URL,
-  type: 'externalLink'
-} as ExternalNavLink];
-
-let defaultSubNavItems:(ExternalNavLink | InternalNavLink | DropdownNavLink | ModalNavLink)[] = [{
-  title: 'About',
-  to: ABOUT_PATH,
-  type: 'internalLink'
-}];
-
-if (process.env.GOOGLE_FORM) {
-  defaultSubNavItems = [...defaultSubNavItems, {
-    title: 'Contact us',
-    src: process.env.GOOGLE_FORM,
-    type: 'modal'
-  }];
-}
-
-const navItems = getNavItemsFromVedaConfig()?.mainNavItems?? defaultMainNavItems;
-const subNavItems = getNavItemsFromVedaConfig()?.subNavItems?? defaultSubNavItems;
-
 function LayoutRoot(props: { children?: ReactNode }) {
   const { children } = props;
 
@@ -101,7 +54,7 @@ function LayoutRoot(props: { children?: ReactNode }) {
         thumbnail={thumbnail}
       />
       {banner && <Banner appTitle={title} {...banner} />}
-      <NavWrapper mainNavItems={navItems} subNavItems={subNavItems} />
+      <NavWrapper mainNavItems={mainNavItems} subNavItems={subNavItems} logo={<Logo />} />
       <PageBody id={PAGE_BODY_ID} tabIndex={-1}>
         <Outlet />
         {children}
